Refactor icon imports across components
Various component files have been updated to import icons from a common icons.tsx file instead of individual icon files. This change centralizes the management of icon resources, making the codebase cleaner and more maintainable. Some icons have also been renamed for consistency.
