+++ src/components/header.tsx
@@ -1,5 +1,5 @@
-import BackIcon from "../icons/back-icon.tsx";
 import {useLocation, useNavigate} from "react-router-dom";
+import {ChevronLeftIcon} from "../icons/icons.tsx";
 
 const Header = () => {
     const location = useLocation()
@@ -9,7 +9,7 @@ const Header = () => {
             <button className={"absolute left-5"}
                     onClick={() => navigate(-1)}
             >
-                <BackIcon/>
+                <ChevronLeftIcon/>
             </button>
             <h2 className={"m-0 font-semibold text-xl"}>workout {location.pathname.split("/")}</h2>
         </div>
