+++ src/icons/icons.tsx
@@ -29,7 +29,7 @@ export const HomeIcon = (props: IconProps) => {
     );
 };
 
-export const BackIcon = (props: IconProps) => {
+export const ChevronLeftIcon = (props: IconProps) => {
     return (
         <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth="1.5"
              stroke="currentColor" className="w-6 h-6"
